+++ src/main/java/com/mindhub/homebanking/models/ClientLoan.java
@@ -12,7 +12,7 @@ public class ClientLoan {
     @GenericGenerator(name = "native", strategy = "native")
     private Long id;
     private double amount;
-    private double payments;
+    private int payments;
 
     @ManyToOne(fetch = FetchType.EAGER)
     @JoinColumn(name="user_id")
@@ -25,7 +25,7 @@ public class ClientLoan {
     public ClientLoan() {
     }
 
-    public ClientLoan(double amount, double payments) {
+    public ClientLoan(double amount, int payments) {
         this.amount = amount;
         this.payments = payments;
     }
@@ -38,7 +38,7 @@ public class ClientLoan {
         return amount;
     }
 
-    public double getPayments() {
+    public int getPayments() {
         return payments;
     }
 
@@ -54,7 +54,7 @@ public class ClientLoan {
         this.amount = amount;
     }
 
-    public void setPayments(double payments) {
+    public void setPayments(int payments) {
         this.payments = payments;
     }
